requirements.txt
faiss-cpu
langchain
langchain-openai
langchain-community
pandas
matplotlib
python-dotenv
streamlit
huggingface-hub
pypdf2
openai
streamlit
chromadb
sentence-transformers
